% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_absolute_path}
\alias{is_absolute_path}
\title{is_absolute_path function from fs}
\usage{
is_absolute_path(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
Re-export of \code{fs::is_absolute_path}.
See the original package documentation for full details.
Test if a path is an absolute path
}
