% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_access}
\alias{is_access}
\title{is_access function from checkmate}
\usage{
is_access(x, access = "")
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{access}{\link{\code{character(1)}}\cr
Single string containing possible characters \sQuote{r}, \sQuote{w} and \sQuote{x} to
force a check for read, write or execute access rights, respectively.
Write and executable rights are not checked on Windows.}
}
\description{
Re-export of \code{checkmate::test_access}.
See the original package documentation for full details.
Check file system access rights
}
