% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_array}
\alias{is_array}
\title{is_array function from checkmate}
\usage{
is_array(
  x,
  mode = NULL,
  any.missing = TRUE,
  d = NULL,
  min.d = NULL,
  max.d = NULL,
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{mode}{\link{\code{character(1)}}\cr
Storage mode of the array. Arrays can hold vectors, i.e. \dQuote{logical},
\dQuote{integer}, \dQuote{integerish}, \dQuote{double}, \dQuote{numeric}, \dQuote{complex},
\dQuote{character} and \dQuote{list}. You can also specify \dQuote{atomic}
here to explicitly prohibit lists. Default is \code{NULL} (no check).
If all values of \code{x} are missing, this check is skipped.}

\item{any.missing}{\link{\code{logical(1)}}\cr
Are missing values allowed? Default is \code{TRUE}.}

\item{d}{\link{\code{integer(1)}}\cr
Exact number of dimensions of array \code{x}.
Default is \code{NULL} (no check).}

\item{min.d}{\link{\code{integer(1)}}\cr
Minimum number of dimensions of array \code{x}.
Default is \code{NULL} (no check).}

\item{max.d}{\link{\code{integer(1)}}\cr
Maximum number of dimensions of array \code{x}.
Default is \code{NULL} (no check).}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_array}.
See the original package documentation for full details.
Check if an argument is an array
}
