% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_atomic}
\alias{is_atomic}
\title{is_atomic function from checkmate}
\usage{
is_atomic(
  x,
  any.missing = TRUE,
  all.missing = TRUE,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  unique = FALSE,
  names = NULL
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{any.missing}{\link{\code{logical(1)}}\cr
Are vectors with missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{\link{\code{logical(1)}}\cr
Are vectors with no non-missing values allowed? Default is \code{TRUE}.
Note that empty vectors do not have non-missing values.}

\item{len}{\link{\code{integer(1)}}\cr
Exact expected length of \code{x}.}

\item{min.len}{\link{\code{integer(1)}}\cr
Minimal length of \code{x}.}

\item{max.len}{\link{\code{integer(1)}}\cr
Maximal length of \code{x}.}

\item{unique}{\link{\code{logical(1)}}\cr
Must all values be unique? Default is \code{FALSE}.}

\item{names}{\link{\code{character(1)}}\cr
Check for names. See \code{\link{checkNamed}} for possible values.
Default is \dQuote{any} which performs no check at all.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}
}
\description{
Re-export of \code{checkmate::test_atomic}.
See the original package documentation for full details.
For the definition of list("atomic"), see list(list("is.atomic")).

Note that \code{NULL} is recognized as a valid atomic value, as in R versions up to version 4.3.x.
For details, see list("https://stat.ethz.ch/pipermail/r-devel/2023-September/082892.html").
}
