% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_attached}
\alias{is_attached}
\title{is_attached function from rlang}
\usage{
is_attached(x)
}
\arguments{
\item{x}{An environment or a search name.}
}
\description{
Re-export of \code{rlang::is_attached}.
See the original package documentation for full details.
The search path is a chain of environments containing exported
functions of attached packages.

The API includes:
list("\n", list(), " ", list(list("base::search()")), " to get the names of environments attached to the\n", "search path.\n", list(), " ", list("search_envs()"), " returns the environments on the search path as a\n", "list.\n", list(), " ", list("pkg_env_name()"), " takes a bare package name and prefixes it with\n", list("\"package:\""), ". Attached package environments have search names of\n", "the form ", list("package:name"), ".\n", list(), " ", list("pkg_env()"), " takes a bare package name and returns the scoped\n",
"environment of packages if they are attached to the search path,\n", "and throws an error otherwise. It is a shortcut for\n", list("search_env(pkg_env_name(\"pkgname\"))"), ".\n", list(), " ", list("global_env()"), " and ", list("base_env()"), " (simple aliases for ", list(list("globalenv()")), "\n", "and ", list(list("baseenv()")), "). These are respectively the first and last\n", "environments of the search path.\n", list(), " ", list("is_attached()"), " returns ", list("TRUE"), " when its argument (a search name\n",
"or a package environment) is attached to the search path.\n")
}
