% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_bare_vector}
\alias{is_bare_vector}
\title{is_bare_vector function from rlang}
\usage{
is_bare_vector(x, n = NULL)
}
\arguments{
\item{x}{Object to be tested.}

\item{n}{Expected length of a vector.}
}
\description{
Re-export of \code{rlang::is_bare_vector}.
See the original package documentation for full details.
These predicates check for a given type but only return list("TRUE") for
bare R objects. Bare objects have no class attributes. For example,
a data frame is a list, but not a bare list.
}
