% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_box}
\alias{is_box}
\title{is_box function from rlang}
\usage{
is_box(x, class = NULL)
}
\arguments{
\item{class}{For \code{new_box()}, an additional class for the
boxed value (in addition to \code{rlang_box}). For \code{is_box()}, a class
or vector of classes passed to \code{\link{inherits_all()}}.}
}
\description{
Re-export of \code{rlang::is_box}.
See the original package documentation for full details.
list("new_box()") is similar to list(list("base::I()")) but it protects a value by
wrapping it in a scalar list rather than by adding an attribute.
list("unbox()") retrieves the boxed value. list("is_box()") tests whether an
object is boxed with optional class. list("as_box()") ensures that a
value is wrapped in a box. list("as_box_if()") does the same but only if
the value matches a predicate.
}
