% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_bytes}
\alias{is_bytes}
\title{is_bytes function from rlang}
\usage{
is_bytes(x, n = NULL)
}
\arguments{
\item{x}{Object to be tested.}

\item{n}{Expected length of a vector.}
}
\description{
Re-export of \code{rlang::is_bytes}.
See the original package documentation for full details.
These type predicates aim to make type testing in R more
consistent. They are wrappers around list(list("base::typeof()")), so operate
at a level beneath S3/S4 etc.
}
