% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_callable}
\alias{is_callable}
\title{is_callable function from rlang}
\usage{
is_callable(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_callable}.
See the original package documentation for full details.
A callable object is an object that can appear in the function
position of a call (as opposed to argument position). This includes
list("symbolic objects") that evaluate to a function or
literal functions embedded in the call.
}
