% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_character}
\alias{is_character}
\alias{are_character,}
\alias{is_not_character,}
\alias{are_not_character}
\alias{is_not_character}
\alias{are_character}
\title{Check if an argument is a vector of type character}
\usage{
is_character(
  x,
  n.chars = NULL,
  min.chars = NULL,
  max.chars = NULL,
  pattern = NULL,
  fixed = NULL,
  ignore.case = FALSE,
  any.missing = TRUE,
  all.missing = TRUE,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  unique = FALSE,
  sorted = FALSE,
  names = NULL,
  typed.missing = FALSE,
  null.ok = FALSE
)

is_not_character(
  x,
  n.chars,
  min.chars,
  max.chars,
  pattern,
  fixed,
  ignore.case,
  any.missing,
  all.missing,
  len,
  min.len,
  max.len,
  unique,
  sorted,
  names,
  typed.missing,
  null.ok
)

are_character(
  x,
  n.chars,
  min.chars,
  max.chars,
  pattern,
  fixed,
  ignore.case,
  any.missing,
  all.missing,
  len,
  min.len,
  max.len,
  unique,
  sorted,
  names,
  typed.missing,
  null.ok
)

are_not_character(
  x,
  n.chars,
  min.chars,
  max.chars,
  pattern,
  fixed,
  ignore.case,
  any.missing,
  all.missing,
  len,
  min.len,
  max.len,
  unique,
  sorted,
  names,
  typed.missing,
  null.ok
)
}
\arguments{
\item{x}{any\cr
Object to check.}

\item{n.chars}{\code{integer(1)}\cr
Exact number of characters for each element of \code{x}.}

\item{min.chars}{\code{integer(1)}\cr
Minimum number of characters for each element of \code{x}.}

\item{max.chars}{\code{integer(1)}\cr
Maximum number of characters for each element of \code{x}.}

\item{pattern}{\code{character(1L)}\cr
Regular expression as used in \code{\link[base]{grepl}}.
All non-missing elements of \code{x} must comply to this pattern.}

\item{fixed}{\code{character(1)}\cr
Substring to detect in \code{x}. Will be used as \code{pattern} in \code{\link[base]{grepl}}
with option \code{fixed} set to \code{TRUE}.
All non-missing elements of \code{x} must contain this substring.}

\item{ignore.case}{\code{logical(1)}\cr
See \code{\link[base]{grepl}}. Default is \code{FALSE}.}

\item{any.missing}{\code{logical(1)}\cr
Are vectors with missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{\code{logical(1)}\cr
Are vectors with no non-missing values allowed? Default is \code{TRUE}.
Note that empty vectors do not have non-missing values.}

\item{len}{\code{integer(1)}\cr
Exact expected length of \code{x}.}

\item{min.len}{\code{integer(1)}\cr
Minimal length of \code{x}.}

\item{max.len}{\code{integer(1)}\cr
Maximal length of \code{x}.}

\item{unique}{\code{logical(1)}\cr
Must all values be unique? Default is \code{FALSE}.}

\item{sorted}{\code{logical(1)}\cr
Elements must be sorted in ascending order. Missing values are ignored.}

\item{names}{\code{character(1)}\cr
Check for names. See \code{\link{checkNamed}} for possible values.
Default is \dQuote{any} which performs no check at all.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}

\item{typed.missing}{\code{logical(1)}\cr
If set to \code{FALSE} (default), all types of missing values (\code{NA}, \code{NA_integer_},
\code{NA_real_}, \code{NA_character_} or \code{NA_character_}) as well as empty vectors are allowed
while type-checking atomic input.
Set to \code{TRUE} to enable strict type checking.}

\item{null.ok}{\code{logical(1)}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\value{
\itemize{
\item Calls to \code{is_character} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_character} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_character(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_character}/\code{are_not_character} negate the output of \code{is_character}/\code{are_character}.}
}
\description{
This is a re-export of \code{\link[checkmate:checkCharacter]{checkmate::test_character()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[checkmate:checkCharacter]{original}} for full details.

To check for scalar strings, see \code{\link{checkString}}.
}
