% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_choice}
\alias{is_choice}
\title{is_choice function from checkmate}
\usage{
is_choice(x, choices, null.ok = FALSE, fmatch = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{choices}{\link{\code{atomic}}\cr
Set of possible values.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}

\item{fmatch}{\link{\code{logical(1)}}\cr
Use the set operations implemented in \code{\link{fmatch}} in package \pkg{fastmatch}.
If \pkg{fastmatch} is not installed, this silently falls back to \code{\link{match}}.
\code{\link{fmatch}} modifies \code{y} by reference:
A hash table is added as attribute which is used in subsequent calls.}
}
\description{
Re-export of \code{checkmate::test_choice}.
See the original package documentation for full details.
Check if an object is an element of a given set
}
