% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_class}
\alias{is_class}
\title{is_class function from checkmate}
\usage{
is_class(x, classes, ordered = FALSE, null.ok = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{classes}{\link{\code{character}}\cr
Class names to check for inheritance with \code{\link{inherits}}.
\code{x} must inherit from all specified classes.}

\item{ordered}{\link{\code{logical(1)}}\cr
Expect \code{x} to be specialized in provided order.
Default is \code{FALSE}.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_class}.
See the original package documentation for full details.
Check the class membership of an argument
}
