% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_condition}
\alias{is_condition}
\title{is_condition function from rlang}
\usage{
is_condition(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_condition}.
See the original package documentation for full details.
Is object a condition?
}
