% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_copyable}
\alias{is_copyable}
\title{is_copyable function from rlang}
\usage{
is_copyable(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_copyable}.
See the original package documentation for full details.
When an object is modified, R generally copies it (sometimes
lazily) to enforce list(list("https://en.wikipedia.org/wiki/Value_semantics"), list("value semantics")).
However, some internal types are uncopyable. If you try to copy
them, either with list("<-") or by argument passing, you actually create
references to the original object rather than actual
copies. Modifying these references can thus have far reaching side
effects.
}
