% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_difftime}
\alias{is_difftime}
\title{is_difftime function from lubridate}
\usage{
is_difftime(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Re-export of \code{lubridate::is.difftime}.
See the original package documentation for full details.
Is x a difftime object?
}
