% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_directory}
\alias{is_directory}
\title{is_directory function from checkmate}
\usage{
is_directory(x, access = "")
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{access}{\link{\code{character(1)}}\cr
Single string containing possible characters \sQuote{r}, \sQuote{w} and \sQuote{x} to
force a check for read, write or execute access rights, respectively.
Write and executable rights are not checked on Windows.}
}
\description{
Re-export of \code{checkmate::test_directory}.
See the original package documentation for full details.
Check for existence and access rights of directories
}
