% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_done_box}
\alias{is_done_box}
\title{is_done_box function from rlang}
\usage{
is_done_box(x, empty = NULL)
}
\arguments{
\item{x}{For \code{done()}, a value to box. For \code{is_done_box()}, a
value to test.}

\item{empty}{Whether the box is empty. If \code{NULL}, \code{is_done_box()}
returns \code{TRUE} for all done boxes. If \code{TRUE}, it returns \code{TRUE}
only for empty boxes. Otherwise it returns \code{TRUE} only for
non-empty boxes.}
}
\description{
Re-export of \code{rlang::is_done_box}.
See the original package documentation for full details.
A value boxed with list("done()") signals to its caller that it
should stop iterating. Use it to shortcircuit a loop.
}
