% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_empty}
\alias{is_empty}
\title{is_empty function from rlang}
\usage{
is_empty(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Re-export of \code{rlang::is_empty}.
See the original package documentation for full details.
Is object an empty vector or NULL?
}
