% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_error}
\alias{is_error}
\title{is_error function from rlang}
\usage{
is_error(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_error}.
See the original package documentation for full details.
Is object a condition?
}
