% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_existing_dir}
\alias{is_existing_dir}
\title{is_existing_dir function from fs}
\usage{
is_existing_dir(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
Re-export of \code{fs::dir_exists}.
See the original package documentation for full details.
list("file_exists(path)") is a shortcut for list("file_access(x, \"exists\")");
list("dir_exists(path)") and list("link_exists(path)") are similar but also check that
the path is a directory or link, respectively. (list("file_exists(path)") returns
list("TRUE") if list("path") exists and it is a directory.)
}
