% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_existing_test_file}
\alias{is_existing_test_file}
\title{is_existing_test_file function from checkmate}
\usage{
is_existing_test_file(x, access = "", extension = NULL)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{access}{\link{\code{character(1)}}\cr
Single string containing possible characters \sQuote{r}, \sQuote{w} and \sQuote{x} to
force a check for read, write or execute access rights, respectively.
Write and executable rights are not checked on Windows.}

\item{extension}{\link{\code{character}}\cr
Vector of allowed file extensions, matched case insensitive.}
}
\description{
Re-export of \code{checkmate::test_file_exists}.
See the original package documentation for full details.
Check existence and access rights of files
}
