% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_flag}
\alias{is_flag}
\title{is_flag function from checkmate}
\usage{
is_flag(x, na.ok = FALSE, null.ok = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{na.ok}{\link{\code{logical(1)}}\cr
Are missing values allowed? Default is \code{FALSE}.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_flag}.
See the original package documentation for full details.
A flag is defined as single logical value.
}
