% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_function}
\alias{is_function}
\title{is_function function from checkmate}
\usage{
is_function(x, args = NULL, ordered = FALSE, nargs = NULL, null.ok = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{args}{\link{\code{character}}\cr
Expected formal arguments. Checks that a function has no arguments if
set to \code{character(0)}.
Default is \code{NULL} (no check).}

\item{ordered}{\link{\code{logical(1)}}\cr
Flag whether the arguments provided in \code{args} must be the first
\code{length(args)} arguments of the function in the specified order.
Default is \code{FALSE}.}

\item{nargs}{\link{\code{integer(1)}}\cr
Required number of arguments, without \code{...}.
Default is \code{NULL} (no check).}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_function}.
See the original package documentation for full details.
Check if an argument is a function
}
