% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_installed}
\alias{is_installed}
\title{is_installed function from rlang}
\usage{
is_installed(pkg, ..., version = NULL, compare = NULL)
}
\arguments{
\item{pkg}{The package names. Can include version requirements,
e.g. \code{"pkg (>= 1.0.0)"}.}

\item{...}{These dots must be empty.}

\item{version}{Minimum versions for \code{pkg}. If supplied, must be the
same length as \code{pkg}. \code{NA} elements stand for any versions.}

\item{compare}{A character vector of comparison operators to use
for \code{version}. If supplied, must be the same length as
\code{version}. If \code{NULL}, \code{>=} is used as default for all
elements. \code{NA} elements in \code{compare} are also set to \code{>=} by
default.}
}
\description{
Re-export of \code{rlang::is_installed}.
See the original package documentation for full details.
These functions check that packages are installed with minimal side
effects. If installed, the packages will be loaded but not
attached.
list("\n", list(), " ", list("is_installed()"), " doesn't interact with the user. It simply\n", "returns ", list("TRUE"), " or ", list("FALSE"), " depending on whether the packages are\n", "installed.\n", list(), " In interactive sessions, ", list("check_installed()"), " asks the user\n", "whether to install missing packages. If the user accepts, the\n", "packages are installed with ", list("pak::pkg_install()"), " if available, or\n", list(list("utils::install.packages()")), " otherwise. If the session is non\n",
"interactive or if the user chooses not to install the packages,\n", "the current evaluation is aborted.\n")

You can disable the prompt by setting the
list("rlib_restart_package_not_found") global option to list("FALSE"). In that
case, missing packages always cause an error.
}
