% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_int}
\alias{is_int}
\title{is_int function from checkmate}
\usage{
is_int(
  x,
  na.ok = FALSE,
  lower = -Inf,
  upper = Inf,
  tol = sqrt(.Machine$double.eps),
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{na.ok}{\link{\code{logical(1)}}\cr
Are missing values allowed? Default is \code{FALSE}.}

\item{lower}{\link{\code{numeric(1)}}\cr
Lower value all elements of \code{x} must be greater than or equal to.}

\item{upper}{\link{\code{numeric(1)}}\cr
Upper value all elements of \code{x} must be lower than or equal to.}

\item{tol}{\link{\code{double(1)}}\cr
Numerical tolerance used to check whether a double or complex can be converted.
Default is \code{sqrt(.Machine$double.eps)}.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_int}.
See the original package documentation for full details.
Check if an argument is a single integerish value
}
