% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_interactive}
\alias{is_interactive}
\title{is_interactive function from rlang}
\usage{
is_interactive()
}
\description{
Re-export of \code{rlang::is_interactive}.
See the original package documentation for full details.
Like list(list("base::interactive()")), list("is_interactive()") returns list("TRUE") when
the function runs interactively and list("FALSE") when it runs in batch
mode. It also checks, in this order:
list("\n", list(), " The ", list("rlang_interactive"), " global option. If set to a single ", list("TRUE"), "\n", "or ", list("FALSE"), ", ", list("is_interactive()"), " returns that value immediately. This\n", "escape hatch is useful in unit tests or to manually turn on\n", "interactive features in RMarkdown outputs.\n", list(), " Whether knitr or testthat is in progress, in which case\n", list("is_interactive()"), " returns ", list("FALSE"), ".\n")

list("with_interactive()") and list("local_interactive()") set the global
option conveniently.
}
