% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_interval}
\alias{is_interval}
\title{is_interval function from lubridate}
\usage{
is_interval(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Re-export of \code{lubridate::is.interval}.
See the original package documentation for full details.
list("interval()") creates an list("Interval") object with the specified start and
end dates. If the start date occurs before the end date, the interval will be
positive. Otherwise, it will be negative. Character vectors in ISO 8601
format are supported from v1.7.2.

list("int_start()")/list("int_end()") and list("int_start<-()")/list("int_end<-()") are
"accessors" and "setters" respectively of the start/end date of an
interval.

list("int_flip()") reverses the order of the start date and end date in an
interval. The new interval takes place during the same timespan as the
original interval, but has the opposite direction.

list("int_shift()") shifts the start and end dates of an interval up or down the
timeline by a specified amount. Note that this may change the exact length of
the interval if the interval is shifted by a Period object. Intervals shifted
by a Duration or difftime object will retain their exact length in seconds.

list("int_overlaps()") tests if two intervals overlap.

list("int_standardize()") ensures all intervals in an interval object are
positive. If an interval is not positive, flip it so that it retains its
endpoints but becomes positive.

list("int_aligns()") tests if two intervals share an endpoint. The direction of
each interval is ignored. int_align tests whether the earliest or latest
moments of each interval occur at the same time.

list("int_diff()") returns the intervals that occur between the elements of a
vector of date-times. list("int_diff()") is similar to the POSIXt and Date
methods of list(list("diff()")), but returns an list("Interval") object instead
of a difftime object.
}
