% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_lang}
\alias{is_lang}
\title{is_lang function from rlang}
\usage{
is_lang(x, name = NULL, n = NULL, ns = NULL)
}
\arguments{
\item{x}{An object to test. Formulas and quosures are treated
literally.}

\item{name}{An optional name that the call should match. It is
passed to \code{\link{sym()}} before matching. This argument is vectorised
and you can supply a vector of names to match. In this case,
\code{is_call()} returns \code{TRUE} if at least one name matches.}

\item{n}{An optional number of arguments that the call should
match.}

\item{ns}{The namespace of the call. If \code{NULL}, the namespace
doesn't participate in the pattern-matching. If an empty string
\code{""} and \code{x} is a namespaced call, \code{is_call()} returns
\code{FALSE}. If any other string, \code{is_call()} checks that \code{x} is
namespaced within \code{ns}.

Can be a character vector of namespaces, in which case the call
has to match at least one of them, otherwise \code{is_call()} returns
\code{FALSE}.}
}
\description{
Re-export of \code{rlang::is_lang}.
See the original package documentation for full details.
list(list("html"), list(list(list("https://lifecycle.r-lib.org/articles/stages.html#deprecated"), list(list(list("lifecycle-deprecated.svg"), list("options: alt='\link{Deprecated}'"))))), list(list("\link{Deprecated}")))
These functions are deprecated, please use list(list("is_call()")) and its list("n")
argument instead.
}
