% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_link}
\alias{is_link}
\title{is_link function from fs}
\usage{
is_link(path)
}
\arguments{
\item{path}{A character vector of one or more paths.}
}
\description{
Re-export of \code{fs::is_link}.
See the original package documentation for full details.
Functions to test for file types
}
