% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_list}
\alias{is_list}
\title{is_list function from checkmate}
\usage{
is_list(
  x,
  types = character(0L),
  any.missing = TRUE,
  all.missing = TRUE,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  unique = FALSE,
  names = NULL,
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{types}{\link{\code{character}}\cr
Character vector of class names. Each list element must inherit
from at least one of the provided types.
The types \dQuote{logical}, \dQuote{integer}, \dQuote{integerish}, \dQuote{double},
\dQuote{numeric}, \dQuote{complex}, \dQuote{character}, \dQuote{factor}, \dQuote{atomic}, \dQuote{vector}
\dQuote{atomicvector}, \dQuote{array}, \dQuote{matrix}, \dQuote{list}, \dQuote{function},
\dQuote{environment} and \dQuote{null} are supported.
For other types \code{\link{inherits}} is used as a fallback to check \code{x}'s inheritance.
Defaults to \code{character(0)} (no check).}

\item{any.missing}{\link{\code{logical(1)}}\cr
Are vectors with missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{\link{\code{logical(1)}}\cr
Are vectors with no non-missing values allowed? Default is \code{TRUE}.
Note that empty vectors do not have non-missing values.}

\item{len}{\link{\code{integer(1)}}\cr
Exact expected length of \code{x}.}

\item{min.len}{\link{\code{integer(1)}}\cr
Minimal length of \code{x}.}

\item{max.len}{\link{\code{integer(1)}}\cr
Maximal length of \code{x}.}

\item{unique}{\link{\code{logical(1)}}\cr
Must all values be unique? Default is \code{FALSE}.}

\item{names}{\link{\code{character(1)}}\cr
Check for names. See \code{\link{checkNamed}} for possible values.
Default is \dQuote{any} which performs no check at all.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_list}.
See the original package documentation for full details.
Check if an argument is a list
}
