% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_message}
\alias{is_message}
\title{is_message function from rlang}
\usage{
is_message(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_message}.
See the original package documentation for full details.
Is object a condition?
}
