% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_missing}
\alias{is_missing}
\title{is_missing function from rlang}
\usage{
is_missing(x)
}
\arguments{
\item{x}{An object that might be the missing argument.}
}
\description{
Re-export of \code{rlang::is_missing}.
See the original package documentation for full details.
These functions help using the missing argument as a regular R
object.
list("\n", list(), " ", list("missing_arg()"), " generates a missing argument.\n", list(), " ", list("is_missing()"), " is like ", list(list("base::missing()")), " but also supports\n", "testing for missing arguments contained in other objects like\n", "lists. It is also more consistent with default arguments which\n", "are never treated as missing (see section below).\n", list(), " ", list("maybe_missing()"), " is useful to pass down an input that might be\n", "missing to another function, potentially substituting by a\n",
"default value. It avoids triggering an \"argument is missing\" error.\n")
}
