% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_multi_class}
\alias{is_multi_class}
\title{is_multi_class function from checkmate}
\usage{
is_multi_class(x, classes, null.ok = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{classes}{\link{\code{character}}\cr
Class names to check for inheritance with \code{\link{inherits}}.
\code{x} must inherit from any of the specified classes.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_multi_class}.
See the original package documentation for full details.
Check the class membership of an argument
}
