% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_named}
\alias{is_named}
\title{is_named function from checkmate}
\usage{
is_named(x, type = "named")
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{type}{\link{character(1)}\cr
Select the check(s) to perform.
\dQuote{unnamed} checks \code{x} to be unnamed.
\dQuote{named} (default) checks \code{x} to be named which excludes names to be \code{NA} or empty (\code{""}).
\dQuote{unique} additionally tests for non-duplicated names.
\dQuote{strict} checks for unique names which comply to R's variable name restrictions.
Note that for zero-length \code{x} every name check evaluates to \code{TRUE}.}
}
\description{
Re-export of \code{checkmate::test_named}.
See the original package documentation for full details.
Check if an argument is named
}
