% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_named2}
\alias{is_named2}
\title{is_named2 function from rlang}
\usage{
is_named2(x)
}
\arguments{
\item{x}{A vector to test.}
}
\description{
Re-export of \code{rlang::is_named2}.
See the original package documentation for full details.
list("\n", list(), " ", list("is_named()"), " is a scalar predicate that checks that ", list("x"), " has a\n", list("names"), " attribute and that none of the names are missing or empty\n", "(", list("NA"), " or ", list("\"\""), ").\n", list(), " ", list("is_named2()"), " is like ", list("is_named()"), " but always returns ", list("TRUE"), " for\n", "empty vectors, even those that don't have a ", list("names"), " attribute.\n", "In other words, it tests for the property that each element of a\n",
"vector is named. ", list("is_named2()"), " composes well with ", list(list("names2()")), "\n", "whereas ", list("is_named()"), " composes with ", list("names()"), ".\n", list(), " ", list("have_name()"), " is a vectorised variant.\n")
}
