% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_names}
\alias{is_names}
\title{is_names function from checkmate}
\usage{
is_names(
  x,
  type = "named",
  subset.of = NULL,
  must.include = NULL,
  permutation.of = NULL,
  identical.to = NULL,
  disjunct.from = NULL,
  what = "names"
)
}
\arguments{
\item{x}{\link{\code{character} || \code{NULL}}\cr
Names to check using rules defined via \code{type}.}

\item{type}{\link{character(1)}\cr
Type of formal check(s) to perform on the names.
\describe{
\item{unnamed:}{Checks \code{x} to be \code{NULL}.}
\item{named:}{Checks \code{x} for regular names which excludes names to be \code{NA} or empty (\code{""}).}
\item{unique:}{Performs checks like with \dQuote{named} and additionally tests for non-duplicated names.}
\item{strict:}{Performs checks like with \dQuote{unique} and additionally fails for names with UTF-8 characters and names which do not comply to R's variable name restrictions.
As regular expression, this is \dQuote{^\link{.}\emph{\link{a-zA-Z}+\link{a-zA-Z0-9._}}$}.}
\item{ids:}{Same as \dQuote{strict}, but does not enforce uniqueness.}
}
Note that for zero-length \code{x}, all these name checks evaluate to \code{TRUE}.}

\item{subset.of}{\link{\code{character}}\cr
Names provided in \code{x} must be subset of the set \code{subset.of}.}

\item{must.include}{\link{\code{character}}\cr
Names provided in \code{x} must be a superset of the set \code{must.include}.}

\item{permutation.of}{\link{\code{character}}\cr
Names provided in \code{x} must be a permutation of the set \code{permutation.of}.
Duplicated names in \code{permutation.of} are stripped out and duplicated names in \code{x}
thus lead to a failed check.
Use this argument instead of \code{identical.to} if the order of the names is not relevant.}

\item{identical.to}{\link{\code{character}}\cr
Names provided in \code{x} must be identical to the vector \code{identical.to}.
Use this argument instead of \code{permutation.of} if the order of the names is relevant.}

\item{disjunct.from}{\link{\code{character}}\cr
Names provided in \code{x} must may not be present in the vector \code{disjunct.from}.}

\item{what}{\link{\code{character(1)}}\cr
Type of name vector to check, e.g. \dQuote{names} (default), \dQuote{colnames} or \dQuote{rownames}.}
}
\description{
Re-export of \code{checkmate::test_names}.
See the original package documentation for full details.
Performs various checks on character vectors, usually names.
}
