% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_node}
\alias{is_node}
\title{is_node function from rlang}
\usage{
is_node(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Re-export of \code{rlang::is_node}.
See the original package documentation for full details.
list("\n", list(), " ", list("is_pairlist()"), " checks that ", list("x"), " has type ", list("pairlist"), ".\n", list(), " ", list("is_node()"), " checks that ", list("x"), " has type ", list("pairlist"), " or ", list("language"), ".\n", "It tests whether ", list("x"), " is a node that has a CAR and a CDR,\n", "including callable nodes (language objects).\n", list(), " ", list("is_node_list()"), " checks that ", list("x"), " has type ", list("pairlist"), " or ", list("NULL"), ".\n", list("NULL"),
" is the empty node list.\n")
}
