% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_number}
\alias{is_number}
\title{is_number function from checkmate}
\usage{
is_number(
  x,
  na.ok = FALSE,
  lower = -Inf,
  upper = Inf,
  finite = FALSE,
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{na.ok}{\link{\code{logical(1)}}\cr
Are missing values allowed? Default is \code{FALSE}.}

\item{lower}{\link{\code{numeric(1)}}\cr
Lower value all elements of \code{x} must be greater than or equal to.}

\item{upper}{\link{\code{numeric(1)}}\cr
Upper value all elements of \code{x} must be lower than or equal to.}

\item{finite}{\link{\code{logical(1)}}\cr
Check for only finite values? Default is \code{FALSE}.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_number}.
See the original package documentation for full details.
Check if an argument is a single numeric value
}
