% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_os}
\alias{is_os}
\title{is_os function from checkmate}
\usage{
is_os(os)
}
\arguments{
\item{os}{\link{\code{character(1)}}\cr
Check the operating system to be in a set with possible elements \dQuote{windows},
\dQuote{mac}, \dQuote{linux} and \dQuote{solaris}.}
}
\description{
Re-export of \code{checkmate::test_os}.
See the original package documentation for full details.
Check the operating system
}
