% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_path_for_output}
\alias{is_path_for_output}
\title{is_path_for_output function from checkmate}
\usage{
is_path_for_output(x, overwrite = FALSE, extension = NULL)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{overwrite}{\link{\code{logical(1)}}\cr
If \code{TRUE}, an existing file in place is allowed if it
it is both readable and writable.
Default is \code{FALSE}.}

\item{extension}{\link{\code{character(1)}}\cr
Extension of the file, e.g. \dQuote{txt} or \dQuote{tar.gz}.}
}
\description{
Re-export of \code{checkmate::test_path_for_output}.
See the original package documentation for full details.
Check if a file path can be used safely to create a file and write to it.

This is checked:
list("\n", " ", list(), list("Does ", list("dirname(x)"), " exist?"), "\n", " ", list(), list("Does no file under path ", list("x"), " exist?"), "\n", " ", list(), list("Is ", list("dirname(x)"), " writable?"), "\n")
Paths are relative to the current working directory.
}
