% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_period}
\alias{is_period}
\title{is_period function from lubridate}
\usage{
is_period(x)
}
\arguments{
\item{x}{Any R object for \code{is.periods} and a numeric value of the number of
units for elementary constructors. With the exception of seconds(), x must
be an integer.}
}
\description{
Re-export of \code{lubridate::is.period}.
See the original package documentation for full details.
list("period()") creates or parses a period object with the specified values.
}
