% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_posixct}
\alias{is_posixct}
\title{is_posixct function from checkmate}
\usage{
is_posixct(
  x,
  lower = NULL,
  upper = NULL,
  any.missing = TRUE,
  all.missing = TRUE,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  unique = FALSE,
  sorted = FALSE,
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{lower}{\link{\code{\link{Date}}}\cr
All non-missing dates in \code{x} must be >= this POSIXct time. Must be provided in the same timezone as \code{x}.}

\item{upper}{\link{\code{\link{Date}}}\cr
All non-missing dates in \code{x} must be <= this POSIXct time. Must be provided in the same timezone as \code{x}.}

\item{any.missing}{\link{\code{logical(1)}}\cr
Are vectors with missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{\link{\code{logical(1)}}\cr
Are vectors with no non-missing values allowed? Default is \code{TRUE}.
Note that empty vectors do not have non-missing values.}

\item{len}{\link{\code{integer(1)}}\cr
Exact expected length of \code{x}.}

\item{min.len}{\link{\code{integer(1)}}\cr
Minimal length of \code{x}.}

\item{max.len}{\link{\code{integer(1)}}\cr
Maximal length of \code{x}.}

\item{unique}{\link{\code{logical(1)}}\cr
Must all values be unique? Default is \code{FALSE}.}

\item{sorted}{\link{\code{logical(1)}}\cr
Elements must be sorted in ascending order. Missing values are ignored.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_posixct}.
See the original package documentation for full details.
Checks that an object is of class list(list("POSIXct")).
}
