% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_POSIXt}
\alias{is_POSIXt}
\title{is_POSIXt function from lubridate}
\usage{
is_POSIXt(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Re-export of \code{lubridate::is.POSIXt}.
See the original package documentation for full details.
list(list("POSIXct()")) mirrors primitive contructors in base R (list(list("double()")),
list(list("character()")) etc.)
}
