% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_primitive_lazy}
\alias{is_primitive_lazy}
\title{is_primitive_lazy function from rlang}
\usage{
is_primitive_lazy(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
Re-export of \code{rlang::is_primitive_lazy}.
See the original package documentation for full details.
The R language defines two different types of functions: primitive
functions, which are low-level, and closures, which are the regular
kind of functions.
}
