% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_quosure}
\alias{is_quosure}
\title{is_quosure function from rlang}
\usage{
is_quosure(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_quosure}.
See the original package documentation for full details.
list("\n", list(), " ", list("new_quosure()"), " wraps any R object (including expressions,\n", "formulas, or other quosures) into a ", list("quosure"), ".\n", list(), " ", list("as_quosure()"), " is similar but it does not rewrap formulas and\n", "quosures.\n")
}
