% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_quosures}
\alias{is_quosures}
\title{is_quosures function from rlang}
\usage{
is_quosures(x)
}
\arguments{
\item{x}{A list of quosures or objects to coerce to quosures.}
}
\description{
Re-export of \code{rlang::is_quosures}.
See the original package documentation for full details.
This small S3 class provides methods for list("[") and list("c()") and ensures
the following invariants:
list("\n", list(), " The list only contains quosures.\n", list(), " It is always named, possibly with a vector of empty strings.\n")

list("new_quosures()") takes a list of quosures and adds the list("quosures")
class and a vector of empty names if needed. list("as_quosures()") calls
list(list("as_quosure()")) on all elements before creating the list("quosures")
object.
}
