% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_r6}
\alias{is_r6}
\title{is_r6 function from checkmate}
\usage{
is_r6(
  x,
  classes = NULL,
  ordered = FALSE,
  cloneable = NULL,
  public = NULL,
  private = NULL,
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{classes}{\link{\code{character}}\cr
Class names to check for inheritance with \code{\link{inherits}}.
\code{x} must inherit from all specified classes.}

\item{ordered}{\link{\code{logical(1)}}\cr
Expect \code{x} to be specialized in provided order.
Default is \code{FALSE}.}

\item{cloneable}{\link{\code{logical(1)}}\cr
If \code{TRUE}, check that \code{x} has a \code{clone} method. If \code{FALSE}, ensure that
\code{x} is not cloneable.}

\item{public}{\link{\code{character}}\cr
Names of expected public slots. This includes active bindings.}

\item{private}{\link{\code{character}}\cr
Names of expected private slots.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_r6}.
See the original package documentation for full details.
Check if an argument is an R6 class
}
