% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_raw}
\alias{is_raw}
\title{is_raw function from checkmate}
\usage{
is_raw(
  x,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  names = NULL,
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{len}{\link{\code{integer(1)}}\cr
Exact expected length of \code{x}.}

\item{min.len}{\link{\code{integer(1)}}\cr
Minimal length of \code{x}.}

\item{max.len}{\link{\code{integer(1)}}\cr
Maximal length of \code{x}.}

\item{names}{\link{\code{character(1)}}\cr
Check for names. See \code{\link{checkNamed}} for possible values.
Default is \dQuote{any} which performs no check at all.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_raw}.
See the original package documentation for full details.
Check if an argument is a raw vector
}
