% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_reference}
\alias{is_reference}
\title{is_reference function from rlang}
\usage{
is_reference(x, y)
}
\description{
Re-export of \code{rlang::is_reference}.
See the original package documentation for full details.
There are typically two situations where two symbols may refer to
the same object.
list("\n", list(), " R objects usually have copy-on-write semantics. This is an\n", "optimisation that ensures that objects are only copied if\n", "needed. When you copy a vector, no memory is actually copied\n", "until you modify either the original object or the copy is\n", "modified.\n", "\n", "Note that the copy-on-write optimisation is an implementation\n", "detail that is not guaranteed by the specification of the R\n", "language.\n", list(), " Assigning an ", list("uncopyable"), " object (like an\n",
"environment) creates a reference. These objects are never copied\n", "even if you modify one of the references.\n")
}
