% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_scalar_na}
\alias{is_scalar_na}
\title{is_scalar_na function from checkmate}
\usage{
is_scalar_na(x, null.ok = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_scalar_na}.
See the original package documentation for full details.
Check if an argument is a single missing value
}
