% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_scoped}
\alias{is_scoped}
\title{is_scoped function from rlang}
\usage{
is_scoped(nm)
}
\arguments{
\item{nm}{The name of an environment attached to the search
path. Call \code{\link{base::search()}} to see what is currently on the path.}
}
\description{
Re-export of \code{rlang::is_scoped}.
See the original package documentation for full details.
list(list("html"), list(list(list("https://lifecycle.r-lib.org/articles/stages.html#deprecated"), list(list(list("lifecycle-deprecated.svg"), list("options: alt='\link{Deprecated}'"))))), list(list("\link{Deprecated}")))

These functions are deprecated as of rlang 0.3.0. Please use
list(list("is_attached()")) instead.
}
