% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_set_equal}
\alias{is_set_equal}
\title{is_set_equal function from checkmate}
\usage{
is_set_equal(x, y, ordered = FALSE, fmatch = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{y}{\link{\code{atomic}}\cr
Set to compare with.}

\item{ordered}{\link{\code{logical(1)}}\cr
Check \code{x} to have the same length and order as \code{y}, i.e.
check using \dQuote{==} while handling \code{NA}s nicely.
Default is \code{FALSE}.}

\item{fmatch}{\link{\code{logical(1)}}\cr
Use the set operations implemented in \code{\link{fmatch}} in package \pkg{fastmatch}.
If \pkg{fastmatch} is not installed, this silently falls back to \code{\link{match}}.
\code{\link{fmatch}} modifies \code{y} by reference:
A hash table is added as attribute which is used in subsequent calls.}
}
\description{
Re-export of \code{checkmate::test_set_equal}.
See the original package documentation for full details.
Check if an argument is equal to a given set
}
