% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_string}
\alias{is_string}
\title{is_string function from checkmate}
\usage{
is_string(
  x,
  na.ok = FALSE,
  n.chars = NULL,
  min.chars = NULL,
  max.chars = NULL,
  pattern = NULL,
  fixed = NULL,
  ignore.case = FALSE,
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{na.ok}{\link{\code{logical(1)}}\cr
Are missing values allowed? Default is \code{FALSE}.}

\item{n.chars}{\link{\code{integer(1)}}\cr
Exact number of characters for each element of \code{x}.}

\item{min.chars}{\link{\code{integer(1)}}\cr
Minimum number of characters for each element of \code{x}.}

\item{max.chars}{\link{\code{integer(1)}}\cr
Maximum number of characters for each element of \code{x}.}

\item{pattern}{\link{\code{character(1L)}}\cr
Regular expression as used in \code{\link{grepl}}.
All non-missing elements of \code{x} must comply to this pattern.}

\item{fixed}{\link{\code{character(1)}}\cr
Substring to detect in \code{x}. Will be used as \code{pattern} in \code{\link{grepl}}
with option \code{fixed} set to \code{TRUE}.
All non-missing elements of \code{x} must contain this substring.}

\item{ignore.case}{\link{\code{logical(1)}}\cr
See \code{\link{grepl}}. Default is \code{FALSE}.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_string}.
See the original package documentation for full details.
A string is defined as a scalar character vector.
To check for vectors of arbitrary length, see list(list("checkCharacter")).
}
