% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_subset}
\alias{is_subset}
\title{is_subset function from checkmate}
\usage{
is_subset(x, choices, empty.ok = TRUE, fmatch = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{choices}{\link{\code{atomic}}\cr
Set of possible values. May be empty.}

\item{empty.ok}{\link{\code{logical(1)}}\cr
Treat zero-length \code{x} as subset of any set \code{choices} (this includes \code{NULL})?
Default is \code{TRUE}.}

\item{fmatch}{\link{\code{logical(1)}}\cr
Use the set operations implemented in \code{\link{fmatch}} in package \pkg{fastmatch}.
If \pkg{fastmatch} is not installed, this silently falls back to \code{\link{match}}.
\code{\link{fmatch}} modifies \code{y} by reference:
A hash table is added as attribute which is used in subsequent calls.}
}
\description{
Re-export of \code{checkmate::test_subset}.
See the original package documentation for full details.
Check if an argument is a subset of a given set
}
