% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_symbol}
\alias{is_symbol}
\title{is_symbol function from rlang}
\usage{
is_symbol(x, name = NULL)
}
\arguments{
\item{x}{An object to test.}

\item{name}{An optional name or vector of names that the symbol
should match.}
}
\description{
Re-export of \code{rlang::is_symbol}.
See the original package documentation for full details.
Is object a symbol?
}
