% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_timepoint}
\alias{is_timepoint}
\title{is_timepoint function from lubridate}
\usage{
is_timepoint(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Re-export of \code{lubridate::is.timepoint}.
See the original package documentation for full details.
An instant is a specific moment in time. Most common date-time
objects (e.g, POSIXct, POSIXlt, and Date objects) are instants.
}
