% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_true}
\alias{is_true}
\title{is_true function from checkmate}
\usage{
is_true(x, na.ok = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{na.ok}{\link{\code{logical(1)}}\cr
Are missing values allowed? Default is \code{FALSE}.}
}
\description{
Re-export of \code{checkmate::test_true}.
See the original package documentation for full details.
Simply checks if an argument is list("TRUE").
}
