% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_warning}
\alias{is_warning}
\title{is_warning function from rlang}
\usage{
is_warning(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_warning}.
See the original package documentation for full details.
Is object a condition?
}
