% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_weakref}
\alias{is_weakref}
\title{is_weakref function from rlang}
\usage{
is_weakref(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_weakref}.
See the original package documentation for full details.
Is object a weak reference?
}
