% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_zap}
\alias{is_zap}
\title{is_zap function from rlang}
\usage{
is_zap(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_zap}.
See the original package documentation for full details.
list("zap()") creates a sentinel object that indicates that an object
should be removed. For instance, named zaps instruct list(list("env_bind()"))
and list(list("call_modify()")) to remove those objects from the environment or
the call.

The advantage of zap objects is that they unambiguously signal the
intent of removing an object. Sentinels like list("NULL") or
list(list("missing_arg()")) are ambiguous because they represent valid R
objects.
}
