% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_access}
\alias{is_access}
\alias{are_access,}
\alias{is_not_access,}
\alias{are_not_access}
\title{Check file system access rights}
\usage{
is_access(...)
}
\arguments{
\item{x}{any\cr
Object to check.}

\item{access}{\code{character(1)}\cr
Single string containing possible characters \sQuote{r}, \sQuote{w} and \sQuote{x} to
force a check for read, write or execute access rights, respectively.
Write and executable rights are not checked on Windows.}
}
\value{
\itemize{
\item Calls to \code{is_access} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_access} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_access(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_access}/\code{are_not_access} negate the output of \code{is_access}/\code{are_access}.
}
}
\description{
Check file system access rights.

This is a re-export of \code{\link[checkmate:checkAccess]{checkmate::test_access()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[checkmate:checkAccess]{original}} for full details.
}
