% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_atomic_vector}
\alias{is_atomic_vector}
\title{is_atomic_vector function from checkmate}
\usage{
is_atomic_vector(
  x,
  any.missing = TRUE,
  all.missing = TRUE,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  unique = FALSE,
  names = NULL
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{any.missing}{\link{\code{logical(1)}}\cr
Are vectors with missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{\link{\code{logical(1)}}\cr
Are vectors with only missing values allowed? Default is \code{TRUE}.}

\item{len}{\link{\code{integer(1)}}\cr
Exact expected length of \code{x}.}

\item{min.len}{\link{\code{integer(1)}}\cr
Minimal length of \code{x}.}

\item{max.len}{\link{\code{integer(1)}}\cr
Maximal length of \code{x}.}

\item{unique}{\link{\code{logical(1)}}\cr
Must all values be unique? Default is \code{FALSE}.}

\item{names}{\link{\code{character(1)}}\cr
Check for names. See \code{\link{checkNamed}} for possible values.
Default is \dQuote{any} which performs no check at all.}
}
\description{
Re-export of \code{checkmate::test_atomic_vector}.
See the original package documentation for full details.
An atomic vector is defined slightly different from specifications in
list(list("is.atomic")) and list(list("is.vector")):
An atomic vector is either list("logical"), list("integer"), list("numeric"),
list("complex"), list("character") or list("raw") and can have any attributes except a
dimension attribute (like matrices).
I.e., a list("factor") is an atomic vector, but a matrix or list("NULL") are not.
In short, this is basically equivalent to list("is.atomic(x) && !is.null(x) && is.null(dim(x))").
}
