% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_bare_environment}
\alias{is_bare_environment}
\title{is_bare_environment function from rlang}
\usage{
is_bare_environment(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Re-export of \code{rlang::is_bare_environment}.
See the original package documentation for full details.
list("is_bare_environment()") tests whether list("x") is an environment without a s3 or
s4 class.
}
