% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_bool}
\alias{is_bool}
\title{is_bool function from rlang}
\usage{
is_bool(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Re-export of \code{rlang::is_bool}.
See the original package documentation for full details.
These predicates check for a given type and whether the vector is
"scalar", that is, of length 1.

In addition to the length check, list("is_string()") and list("is_bool()")
return list("FALSE") if their input is missing. This is useful for
type-checking arguments, when your function expects a single string
or a single list("TRUE") or list("FALSE").
}
