% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_choice}
\alias{is_choice}
\alias{are_choice,}
\alias{is_not_choice,}
\alias{are_not_choice}
\alias{is_not_choice}
\alias{are_choice}
\title{Check if an object is an element of a given set}
\usage{
is_choice(x, choices, null.ok = FALSE, fmatch = FALSE)

is_not_choice(x, choices, null.ok, fmatch)

are_choice(x, choices, null.ok, fmatch)

are_not_choice(x, choices, null.ok, fmatch)
}
\arguments{
\item{x}{any\cr
Object to check.}

\item{choices}{\code{atomic}\cr
Set of possible values.}

\item{null.ok}{\code{logical(1)}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}

\item{fmatch}{\code{logical(1)}\cr
Use the set operations implemented in \code{\link[fastmatch]{fmatch}} in package \pkg{fastmatch}.
If \pkg{fastmatch} is not installed, this silently falls back to \code{\link[base]{match}}.
\code{\link[fastmatch]{fmatch}} modifies \code{y} by reference:
A hash table is added as attribute which is used in subsequent calls.}
}
\value{
\itemize{
\item Calls to \code{is_choice} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_choice} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_choice(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_choice}/\code{are_not_choice} negate the output of \code{is_choice}/\code{are_choice}.}
}
\description{
This is a re-export of \code{\link[checkmate:checkChoice]{checkmate::test_choice()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[checkmate:checkChoice]{original}} for full details.

Check if an object is an element of a given set.
}
