% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_class}
\alias{is_class}
\alias{are_class,}
\alias{is_not_class,}
\alias{are_not_class}
\alias{is_not_class}
\alias{are_class}
\title{Check the class membership of an argument}
\usage{
is_class(x, classes, ordered = FALSE, null.ok = FALSE)

is_not_class(x, classes, ordered, null.ok)

are_class(x, classes, ordered, null.ok)

are_not_class(x, classes, ordered, null.ok)
}
\arguments{
\item{x}{any\cr
Object to check.}

\item{classes}{\code{character}\cr
Class names to check for inheritance with \code{\link[base]{inherits}}.
\code{x} must inherit from all specified classes.}

\item{ordered}{\code{logical(1)}\cr
Expect \code{x} to be specialized in provided order.
Default is \code{FALSE}.}

\item{null.ok}{\code{logical(1)}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\value{
\itemize{
\item Calls to \code{is_class} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_class} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_class(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_class}/\code{are_not_class} negate the output of \code{is_class}/\code{are_class}.}
}
\description{
This is a re-export of \code{\link[checkmate:checkClass]{checkmate::test_class()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[checkmate:checkClass]{original}} for full details.

Check the class membership of an argument.
}
