% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_count}
\alias{is_count}
\title{is_count function from checkmate}
\usage{
is_count(
  x,
  na.ok = FALSE,
  positive = FALSE,
  tol = sqrt(.Machine$double.eps),
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{na.ok}{\link{\code{logical(1)}}\cr
Are missing values allowed? Default is \code{FALSE}.}

\item{positive}{\link{\code{logical(1)}}\cr
Must \code{x} be positive (>= 1)?
Default is \code{FALSE}, allowing 0.}

\item{tol}{\link{\code{double(1)}}\cr
Numerical tolerance used to check whether a double or complex can be converted.
Default is \code{sqrt(.Machine$double.eps)}.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_count}.
See the original package documentation for full details.
A count is defined as non-negative integerish value.
}
