% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_dictionaryish}
\alias{is_dictionaryish}
\title{is_dictionaryish function from rlang}
\usage{
is_dictionaryish(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Re-export of \code{rlang::is_dictionaryish}.
See the original package documentation for full details.
Like list(list("is_named()")) but also checks that names are unique.
}
