% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_disjunct}
\alias{is_disjunct}
\title{is_disjunct function from checkmate}
\usage{
is_disjunct(x, y, fmatch = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{y}{\link{\code{atomic}}\cr
Other Set.}

\item{fmatch}{\link{\code{logical(1)}}\cr
Use the set operations implemented in \code{\link{fmatch}} in package \pkg{fastmatch}.
If \pkg{fastmatch} is not installed, this silently falls back to \code{\link{match}}.
\code{\link{fmatch}} modifies \code{y} by reference:
A hash table is added as attribute which is used in subsequent calls.}
}
\description{
Re-export of \code{checkmate::test_disjunct}.
See the original package documentation for full details.
Check if an argument is disjunct from a given set
}
