% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_done_box}
\alias{is_done_box}
\alias{are_done_box,}
\alias{is_not_done_box,}
\alias{are_not_done_box}
\alias{is_not_done_box}
\alias{are_done_box}
\title{Box a final value for early termination}
\usage{
is_done_box(x, empty = NULL)

is_not_done_box(x, empty = NULL)

are_done_box(x, empty = NULL)

are_not_done_box(x, empty = NULL)
}
\arguments{
\item{x}{For \code{done()}, a value to box. For \code{is_done_box()}, a
value to test.}

\item{empty}{Whether the box is empty. If \code{NULL}, \code{is_done_box()}
returns \code{TRUE} for all done boxes. If \code{TRUE}, it returns \code{TRUE}
only for empty boxes. Otherwise it returns \code{TRUE} only for
non-empty boxes.}
}
\value{
\itemize{
\item Calls to \code{is_done_box} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_done_box} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_done_box(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_done_box}/\code{are_not_done_box} negate the output of \code{is_done_box}/\code{are_done_box}.}
}
\description{
This is a re-export of \code{\link[rlang:done]{rlang::is_done_box()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[rlang:done]{original}} for full details.

A value boxed with \code{done()} signals to its caller that it
should stop iterating. Use it to shortcircuit a loop.
}
