% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_duration}
\alias{is_duration}
\title{is_duration function from lubridate}
\usage{
is_duration(x)
}
\arguments{
\item{x}{numeric value of the number of units to be contained in the
duration.}
}
\description{
Re-export of \code{lubridate::is.duration}.
See the original package documentation for full details.
list("duration()") creates a duration object with the specified values. Entries
for different units are cumulative. durations display as the number of
seconds in a time span. When this number is large, durations also display an
estimate in larger units, however, the underlying object is always recorded
as a fixed number of seconds. For display and creation purposes, units are
converted to seconds using their most common lengths in seconds. Minutes = 60
seconds, hours = 3600 seconds, days = 86400 seconds, weeks = 604800. Units
larger than weeks are not used due to their variability.
}
