% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_environment}
\alias{is_environment}
\title{is_environment function from checkmate}
\usage{
is_environment(x, contains = character(0L), null.ok = FALSE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{contains}{\link{\code{character}}\cr
Vector of object names expected in the environment.
Defaults to \code{character(0)}.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_environment}.
See the original package documentation for full details.
Check if an argument is an environment
}
