% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_file_empty}
\alias{is_file_empty}
\title{is_file_empty function from fs}
\usage{
is_file_empty(path, follow = TRUE)
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{follow}{If \code{TRUE}, symbolic links will be followed (recursively) and
the results will be that of the final file rather than the link.}
}
\description{
Re-export of \code{fs::is_file_empty}.
See the original package documentation for full details.
Functions to test for file types
}
