% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_list}
\alias{is_list}
\alias{are_list,}
\alias{is_not_list,}
\alias{are_not_list}
\alias{is_not_list}
\alias{are_list}
\title{Check if an argument is a list}
\usage{
is_list(
  x,
  types = character(0L),
  any.missing = TRUE,
  all.missing = TRUE,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  unique = FALSE,
  names = NULL,
  null.ok = FALSE
)

is_not_list(
  x,
  types,
  any.missing,
  all.missing,
  len,
  min.len,
  max.len,
  unique,
  names,
  null.ok
)

are_list(
  x,
  types,
  any.missing,
  all.missing,
  len,
  min.len,
  max.len,
  unique,
  names,
  null.ok
)

are_not_list(
  x,
  types,
  any.missing,
  all.missing,
  len,
  min.len,
  max.len,
  unique,
  names,
  null.ok
)
}
\arguments{
\item{x}{any\cr
Object to check.}

\item{types}{\code{character}\cr
Character vector of class names. Each list element must inherit
from at least one of the provided types.
The types \dQuote{logical}, \dQuote{integer}, \dQuote{integerish}, \dQuote{double},
\dQuote{numeric}, \dQuote{complex}, \dQuote{character}, \dQuote{factor}, \dQuote{atomic}, \dQuote{vector}
\dQuote{atomicvector}, \dQuote{array}, \dQuote{matrix}, \dQuote{list}, \dQuote{function},
\dQuote{environment} and \dQuote{null} are supported.
For other types \code{\link{inherits}} is used as a fallback to check \code{x}'s inheritance.
Defaults to \code{character(0)} (no check).}

\item{any.missing}{\code{logical(1)}\cr
Are vectors with missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{\code{logical(1)}\cr
Are vectors with no non-missing values allowed? Default is \code{TRUE}.
Note that empty vectors do not have non-missing values.}

\item{len}{\code{integer(1)}\cr
Exact expected length of \code{x}.}

\item{min.len}{\code{integer(1)}\cr
Minimal length of \code{x}.}

\item{max.len}{\code{integer(1)}\cr
Maximal length of \code{x}.}

\item{unique}{\code{logical(1)}\cr
Must all values be unique? Default is \code{FALSE}.}

\item{names}{\code{character(1)}\cr
Check for names. See \code{\link{checkNamed}} for possible values.
Default is \dQuote{any} which performs no check at all.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}

\item{null.ok}{\code{logical(1)}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\value{
\itemize{
\item Calls to \code{is_list} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_list} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_list(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_list}/\code{are_not_list} negate the output of \code{is_list}/\code{are_list}.
}
}
\description{
This is a re-export of \code{\link[checkmate:checkList]{checkmate::test_list()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[checkmate:checkList]{original}} for full details.

Check if an argument is a list.
}
