% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_matrix}
\alias{is_matrix}
\title{is_matrix function from checkmate}
\usage{
is_matrix(
  x,
  mode = NULL,
  any.missing = TRUE,
  all.missing = TRUE,
  min.rows = NULL,
  max.rows = NULL,
  min.cols = NULL,
  max.cols = NULL,
  nrows = NULL,
  ncols = NULL,
  row.names = NULL,
  col.names = NULL,
  null.ok = FALSE
)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{mode}{\link{\code{character(1)}}\cr
Storage mode of the array. Arrays can hold vectors, i.e. \dQuote{logical},
\dQuote{integer}, \dQuote{integerish}, \dQuote{double}, \dQuote{numeric}, \dQuote{complex},
\dQuote{character} and \dQuote{list}. You can also specify \dQuote{atomic}
here to explicitly prohibit lists. Default is \code{NULL} (no check).
If all values of \code{x} are missing, this check is skipped.}

\item{any.missing}{\link{\code{logical(1)}}\cr
Are missing values allowed? Default is \code{TRUE}.}

\item{all.missing}{\link{\code{logical(1)}}\cr
Are matrices with only missing values allowed? Default is \code{TRUE}.}

\item{min.rows}{\link{\code{integer(1)}}\cr
Minimum number of rows.}

\item{max.rows}{\link{\code{integer(1)}}\cr
Maximum number of rows.}

\item{min.cols}{\link{\code{integer(1)}}\cr
Minimum number of columns.}

\item{max.cols}{\link{\code{integer(1)}}\cr
Maximum number of columns.}

\item{nrows}{\link{\code{integer(1)}}\cr
Exact number of rows.}

\item{ncols}{\link{\code{integer(1)}}\cr
Exact number of columns.}

\item{row.names}{\link{\code{character(1)}}\cr
Check for row names. Default is \dQuote{NULL} (no check).
See \code{\link{checkNamed}} for possible values.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}

\item{col.names}{\link{\code{character(1)}}\cr
Check for column names. Default is \dQuote{NULL} (no check).
See \code{\link{checkNamed}} for possible values.
Note that you can use \code{\link{checkSubset}} to test for a specific set of names.}

\item{null.ok}{\link{\code{logical(1)}}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\description{
Re-export of \code{checkmate::test_matrix}.
See the original package documentation for full details.
Check if an argument is a matrix
}
