% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_namespace}
\alias{is_namespace}
\title{is_namespace function from rlang}
\usage{
is_namespace(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Re-export of \code{rlang::is_namespace}.
See the original package documentation for full details.
Is an object a namespace environment?
}
