% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_null}
\alias{is_null}
\title{is_null function from checkmate}
\usage{
is_null(x)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}
}
\description{
Re-export of \code{checkmate::test_null}.
See the original package documentation for full details.
Check if an argument is NULL
}
