% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_path_for_output}
\alias{is_path_for_output}
\alias{are_path_for_output,}
\alias{is_not_path_for_output,}
\alias{are_not_path_for_output}
\alias{is_not_path_for_output}
\alias{are_path_for_output}
\title{Check if a path is suited for creating an output file}
\usage{
is_path_for_output(x, overwrite = FALSE, extension = NULL)

is_not_path_for_output(x, overwrite, extension)

are_path_for_output(x, overwrite, extension)

are_not_path_for_output(x, overwrite, extension)
}
\arguments{
\item{x}{any\cr
Object to check.}

\item{overwrite}{\code{logical(1)}\cr
If \code{TRUE}, an existing file in place is allowed if it
it is both readable and writable.
Default is \code{FALSE}.}

\item{extension}{\code{character(1)}\cr
Extension of the file, e.g. \dQuote{txt} or \dQuote{tar.gz}.}
}
\value{
\itemize{
\item Calls to \code{is_path_for_output} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_path_for_output} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_path_for_output(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_path_for_output}/\code{are_not_path_for_output} negate the output of \code{is_path_for_output}/\code{are_path_for_output}.}
}
\description{
This is a re-export of \code{\link[checkmate:checkPathForOutput]{checkmate::test_path_for_output()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[checkmate:checkPathForOutput]{original}} for full details.

Check if a file path can be used safely to create a file and write to it.

This is checked:
\itemize{
 \item{Does list("dirname(x)") exist?
 }\item{Does no file under path list("x") exist?
 }\item{Is list("dirname(x)") writable?
}}
Paths are relative to the current working directory.
}
