% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_permutation}
\alias{is_permutation}
\title{is_permutation function from checkmate}
\usage{
is_permutation(x, y, na.ok = TRUE)
}
\arguments{
\item{x}{\link{any}\cr
Object to check.}

\item{y}{\link{\code{atomic}}\cr
Vector to compare with. Atomic vector of type other than raw.}

\item{na.ok}{\link{\code{logical(1)}}\cr
Are missing values allowed? Default is \code{TRUE}.}
}
\description{
Re-export of \code{checkmate::test_permutation}.
See the original package documentation for full details.
In contrast to list(list("checkSetEqual")), the function tests for a true
permutation of the two vectors and also considers duplicated values.
Missing values are being treated as actual values by default.
Does not work on raw values.
}
