% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_POSIXlt}
\alias{is_POSIXlt}
\alias{are_POSIXlt,}
\alias{is_not_POSIXlt,}
\alias{are_not_POSIXlt}
\alias{is_not_POSIXlt}
\alias{are_POSIXlt}
\title{Various POSIX utilities}
\usage{
is_POSIXlt(x)

is_not_POSIXlt(x)

are_POSIXlt(x)

are_not_POSIXlt(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\itemize{
\item Calls to \code{is_POSIXlt} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_POSIXlt} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_POSIXlt(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_POSIXlt}/\code{are_not_POSIXlt} negate the output of \code{is_POSIXlt}/\code{are_POSIXlt}.}
}
\description{
This is a re-export of \code{\link[lubridate:posix_utils]{lubridate::is.POSIXlt()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[lubridate:posix_utils]{original}} for full details.

\code{\link[lubridate:POSIXct]{POSIXct()}} mirrors primitive contructors in base R (\code{\link[=double]{double()}},
\code{\link[=character]{character()}} etc.).
}
