% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_POSIXt}
\alias{is_POSIXt}
\alias{are_POSIXt,}
\alias{is_not_POSIXt,}
\alias{are_not_POSIXt}
\alias{is_not_POSIXt}
\alias{are_POSIXt}
\title{Various POSIX utilities}
\usage{
is_POSIXt(x)

is_not_POSIXt(x)

are_POSIXt(x)

are_not_POSIXt(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\itemize{
\item Calls to \code{is_POSIXt} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_POSIXt} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_POSIXt(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_POSIXt}/\code{are_not_POSIXt} negate the output of \code{is_POSIXt}/\code{are_POSIXt}.
}
}
\description{
This is a re-export of \code{\link[lubridate:posix_utils]{lubridate::is.POSIXt()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[lubridate:posix_utils]{original}} for full details.

\code{\link{POSIXct()}} mirrors primitive contructors in base R (\code{\link{double()}},
\code{\link{character()}} etc.).
}
