% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_quosures}
\alias{is_quosures}
\alias{are_quosures,}
\alias{is_not_quosures,}
\alias{are_not_quosures}
\title{Create a list of quosures}
\usage{
is_quosures(...)
}
\arguments{
\item{x}{A list of quosures or objects to coerce to quosures.}
}
\value{
\itemize{
\item Calls to \code{is_quosures} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_quosures} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_quosures(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_quosures}/\code{are_not_quosures} negate the output of \code{is_quosures}/\code{are_quosures}.
}
}
\description{
This small S3 class provides methods for \code{[} and \code{c()} and ensures the following invariants: \itemize{ \item The list only contains quosures. \item It is always named, possibly with a vector of empty strings. } \code{new_quosures()} takes a list of quosures and adds the \code{quosures} class and a vector of empty names if needed. \code{as_quosures()} calls \code{\link{as_quosure()}} on all elements before creating the \code{quosures} object.

This is a re-export of \code{\link[rlang:new_quosures]{rlang::is_quosures()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[rlang:new_quosures]{original}} for full details.
}
