% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_raw}
\alias{is_raw}
\alias{are_raw,}
\alias{is_not_raw,}
\alias{are_not_raw}
\alias{is_not_raw}
\alias{are_raw}
\title{Check if an argument is a raw vector}
\usage{
is_raw(
  x,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  names = NULL,
  null.ok = FALSE
)

is_not_raw(
  x,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  names = NULL,
  null.ok = FALSE
)

are_raw(
  x,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  names = NULL,
  null.ok = FALSE
)

are_not_raw(
  x,
  len = NULL,
  min.len = NULL,
  max.len = NULL,
  names = NULL,
  null.ok = FALSE
)
}
\arguments{
\item{x}{any\cr
Object to check.}

\item{len}{\code{integer(1)}\cr
Exact expected length of \code{x}.}

\item{min.len}{\code{integer(1)}\cr
Minimal length of \code{x}.}

\item{max.len}{\code{integer(1)}\cr
Maximal length of \code{x}.}

\item{names}{\code{character(1)}\cr
Check for names. See \code{\link{checkNamed}} for possible values.
Default is \dQuote{any} which performs no check at all.
Note that you can use \code{\link{checkSubset}} to check for a specific set of names.}

\item{null.ok}{\code{logical(1)}\cr
If set to \code{TRUE}, \code{x} may also be \code{NULL}.
In this case only a type check of \code{x} is performed, all additional checks are disabled.}
}
\value{
\itemize{
\item Calls to \code{is_raw} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_raw} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_raw(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_raw}/\code{are_not_raw} negate the output of \code{is_raw}/\code{are_raw}.}
}
\description{
This is a re-export of \code{\link[checkmate:checkRaw]{checkmate::test_raw()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[checkmate:checkRaw]{original}} for full details.

Check if an argument is a raw vector.
}
