% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_symbol}
\alias{is_symbol}
\alias{are_symbol,}
\alias{is_not_symbol,}
\alias{are_not_symbol}
\title{Is object a symbol?}
\usage{
is_symbol(...)
}
\arguments{
\item{x}{An object to test.}

\item{name}{An optional name or vector of names that the symbol
should match.}
}
\value{
\itemize{
\item Calls to \code{is_symbol} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_symbol} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_symbol(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_symbol}/\code{are_not_symbol} negate the output of \code{is_symbol}/\code{are_symbol}.
}
}
\description{
Is object a symbol?.

This is a re-export of \code{\link[rlang:is_symbol]{rlang::is_symbol()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[rlang:is_symbol]{original}} for full details.
}
