% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_timepoint}
\alias{is_timepoint}
\alias{are_timepoint,}
\alias{is_not_timepoint,}
\alias{are_not_timepoint}
\alias{is_not_timepoint}
\alias{are_timepoint}
\title{Is x a date-time object?}
\usage{
is_timepoint(x)

is_not_timepoint(x)

are_timepoint(x)

are_not_timepoint(x)
}
\arguments{
\item{x}{an R object}
}
\value{
\itemize{
\item Calls to \code{is_timepoint} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_timepoint} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_timepoint(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_timepoint}/\code{are_not_timepoint} negate the output of \code{is_timepoint}/\code{are_timepoint}.
}
}
\description{
This is a re-export of \code{\link[lubridate:is.instant]{lubridate::is.timepoint()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[lubridate:is.instant]{original}} for full details.

An instant is a specific moment in time. Most common date-time
objects (e.g, POSIXct, POSIXlt, and Date objects) are instants.
}
