% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_timespan}
\alias{is_timespan}
\title{is_timespan function from lubridate}
\usage{
is_timespan(x)
}
\arguments{
\item{x}{an R object}
}
\description{
Re-export of \code{lubridate::is.timespan}.
See the original package documentation for full details.
Is x a length of time?
}
