% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_warning}
\alias{is_warning}
\alias{are_warning,}
\alias{is_not_warning,}
\alias{are_not_warning}
\title{Is object a condition?}
\usage{
is_warning(...)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\itemize{
\item Calls to \code{is_warning} are guaranteed to return a scalar boolean (ie. a single \code{TRUE} or \code{FALSE} value). If an argument of length > 1 is given, \code{FALSE} is returned.
\item \code{are_warning} is a wrapper around \code{\link[purrr]{map_lgl}(vec, \(i) is_warning(i, ...))}. A boolean vector of the same length as the input is guaranteed.
\item Calls to \code{is_not_warning}/\code{are_not_warning} negate the output of \code{is_warning}/\code{are_warning}.
}
}
\description{
Is object a condition?.

This is a re-export of \code{\link[rlang:is_condition]{rlang::is_warning()}}, modified to have standardised naming and standardised vector handling.
Documentation is atuomatically generated from the original package documentation. See the \code{\link[rlang:is_condition]{original}} for full details.
}
